+++ src/pages/Senpai/TodoList/components/TodoForm.jsx
@@ -1,6 +1,6 @@
 import clsx from 'clsx';
 import { useState } from 'react';
-import Button from './Button_test';
+import Button from './Button';
 
 function TodoForm({ item, updateTodo, toggleDataState, deleteItem }) {
   const [inputValue, setInputValue] = useState(item.text);
